+++ src/hexsample/cli.py
@@ -114,6 +114,15 @@ class CliArgumentParser(argparse.ArgumentParser):
         self.add_logging_level(quicklook)
         quicklook.set_defaults(runner=pipeline.quicklook)
 
+        # Convert a .mdat3 file to a HDF5 digi file?
+        convert = subparsers.add_parser("convert",
+            help="convert a .mdat3 file to a HDF5 digi file",
+            formatter_class=self._FORMATTER_CLASS)
+        self.add_input_file(convert)
+        self.add_num_events(convert, default=None, intent="converted")
+        self.add_logging_level(convert)
+        convert.set_defaults(runner=pipeline.mdat3_to_digi)
+
     @staticmethod
     def add_input_file(parser: argparse.ArgumentParser) -> None:
         """Add an option for the input file.