+++ docs/release_notes.rst
@@ -3,13 +3,17 @@
 Release notes
 =============
 
+* New `legacy.py` module containing the code to convert .mdat3 files to .h5 files.
+* New command `convert` to convert .mdat3 files to .h5 files using the CLI.
 * Added an actual interactive event display with some GUI.
 * Added support for random access in event files.
 * Small cleanup in the fileio.
 * Pull requests merged and issues closed:
 
   - https://github.com/lucabaldini/hexsample/pull/103
+  - https://github.com/lucabaldini/hexsample/pull/100
   - https://github.com/lucabaldini/hexsample/issues/93
+  - https://github.com/lucabaldini/hexsample/issues/67
 
 
 Version 0.13.3 (2026-02-10)